fix: improve error handling and code quality
- Fix download operation to properly handle None data on errors
- Add comprehensive unit tests for error detection and reporting
- Verify 100% failure rate with bad configurations
- Fix type annotations for mypy compliance
- Clean up unused imports detected by ruff
- Format code with ruff formatter

All quality checks passing:
- ruff linting: no errors
- mypy type checking: no issues
- pytest unit tests: 27 passed

+++ src/chopsticks/metrics/models.py
@@ -102,7 +102,7 @@ class AggregatedMetrics:
     operations: Dict[str, Any]
     duration_ms: StatisticalSummary
     throughput_mbps: StatisticalSummary
-    object_size_bytes: Dict[str, int]
+    object_size_bytes: Dict[str, float]
     request_rate: Dict[str, float]
 
     def to_dict(self) -> Dict[str, Any]:
